Add DownloadRepository as internal class
